Enable dry run and syntax check.

+++ examples/cicd/bin/build-archive.sh
@@ -11,11 +11,14 @@ helpAndExit() {
   echo "${CMD_NAME}  -  build and configure YAMLES archive"
   echo ""
   echo "Usage:"
-  echo "  ${CMD_NAME} -e ENV [--debug]"
+  echo "  ${CMD_NAME} -e ENV [--mode MODE] [--debug]"
   echo ""
   echo "Options:"
   echo "  -e ENV"
   echo "      Target environment (local, test or prod)."
+  echo ""
+  echo "  --mode MODE"
+  echo "      Execution mode (CONFIG, DRY_RUN, SYNTAX_CHECK)"
   echo ""
   echo "  --debug"
   echo "      Enable debug messages"
@@ -28,6 +31,7 @@ helpAndExit() {
 
 parseArgs() {
   DEBUG=0
+  MODE="CONFIG"
   while [[ $# -gt 0 ]]
   do
     option="$1"
@@ -50,6 +54,17 @@ parseArgs() {
           esac
         else
           echo "ERROR: missing environment"
+          exit 1
+        fi
+        ;;
+
+      --mode)
+        if [[ $# -gt 1 ]]
+        then
+          MODE="$2"
+          shift 2
+        else
+          echo "ERROR: missing mode"
           exit 1
         fi
         ;;
@@ -210,12 +225,12 @@ echo ""
 echo "============================================================"
 echo "== Configure project - certificates"
 echo "============================================================"
-args=("--audit=${TARGET_TMP_DIR}/audit.log" "merge" "certs" "--project=${TMP_PRJ}")
+args=("--audit=${TARGET_TMP_DIR}/audit.log" "merge" "--mode=${MODE}" "certs" "--project=${TMP_PRJ}")
 case $OPT_ENV in
   local)
     args+=( \
       # Configure certificates for local development environment
-      "--config=${CFG_DIR}/devs/local/certificates.yaml" \
+      "--certs=${CFG_DIR}/devs/local/certificates.yaml" \
 
       # Use local lookups for passphrases
       "--lookup-functions=${CFG_DIR}/devs/local/lookup-func.yaml" \
@@ -224,7 +239,7 @@ case $OPT_ENV in
   test)
     args+=( \
       # Configure test certificates provided by operators
-      "--config=${CFG_DIR}/ops/test/certificates.yaml" \
+      "--certs=${CFG_DIR}/ops/test/certificates.yaml" \
 
       # Use ops-test lookups for secrets
       "--lookup-functions=${CFG_DIR}/ops/test/lookup-func.yaml" \
@@ -233,7 +248,7 @@ case $OPT_ENV in
   prod)
     args+=( \
       # Configure production certificates provided by operators
-      "--config=${CFG_DIR}/ops/prod/certificates.yaml" \
+      "--certs=${CFG_DIR}/ops/prod/certificates.yaml" \
 
       # Use ops-prod lookups for secrets
       "--lookup-functions=${CFG_DIR}/ops/prod/lookup-func.yaml" \
@@ -250,16 +265,16 @@ echo ""
 echo "============================================================"
 echo "== Configure project - values"
 echo "============================================================"
-args=( "--audit=${TARGET_TMP_DIR}/audit.log" "merge" "config" "--project=${TMP_PRJ}")
+args=( "--audit=${TARGET_TMP_DIR}/audit.log" "merge" "--mode=${MODE}" "config" "--project=${TMP_PRJ}")
 case $OPT_ENV in
   local)
     args+=( \
       # Configure general settings
-      "--config=${CFG_DIR}/devs/local/values.yaml" \
+      "--fragment=${CFG_DIR}/devs/local/values.yaml" \
 
       # Configure values for which only developers are responsible for
       # in a separate file.
-      "--config=${CFG_DIR}/devs/local/devs-values.yaml" \
+      "--fragment=${CFG_DIR}/devs/local/devs-values.yaml" \
 
       # Configure lookup functions
       "--lookup-functions=${CFG_DIR}/devs/local/lookup-func.yaml" \
@@ -268,15 +283,15 @@ case $OPT_ENV in
   test)
     args+=( \
       # Read values for which only developers are responsible for
-      "--config=${CFG_DIR}/devs/test/devs-values.yaml" \
+      "--fragment=${CFG_DIR}/devs/test/devs-values.yaml" \
 
       # Configure lookup functions (developers)
       "--lookup-functions=${CFG_DIR}/devs/test/lookup-func.yaml" \
 
       # Read values for which operators are responsible for
       # (must be after the developers to configuration to prevent overwrite)
-      "--config=${CFG_DIR}/ops/all/values.yaml" \
-      "--config=${CFG_DIR}/ops/test/values.yaml" \
+      "--fragment=${CFG_DIR}/ops/all/values.yaml" \
+      "--fragment=${CFG_DIR}/ops/test/values.yaml" \
 
       # Configure lookup functions (ops-test)
       "--lookup-functions=${CFG_DIR}/ops/test/lookup-func.yaml" \
@@ -285,15 +300,15 @@ case $OPT_ENV in
   prod)
     args+=( \
       # Read values for which only developers are responsible for    
-      "--config=${CFG_DIR}/devs/prod/devs-values.yaml" \
+      "--fragment=${CFG_DIR}/devs/prod/devs-values.yaml" \
 
       # Configure lookup functions (developers)
       "--lookup-functions=${CFG_DIR}/devs/prod/lookup-func.yaml" \
 
       # Read values for which operators are responsible for
       # (must be after the developers to configuration to prevent overwrite)
-      "--config=${CFG_DIR}/ops/all/values.yaml" \
-      "--config=${CFG_DIR}/ops/prod/values.yaml" \
+      "--fragment=${CFG_DIR}/ops/all/values.yaml" \
+      "--fragment=${CFG_DIR}/ops/prod/values.yaml" \
 
       # Configure lookup functions (ops-prod)
       "--lookup-functions=${CFG_DIR}/ops/prod/lookup-func.yaml" \
@@ -305,31 +320,33 @@ case $OPT_ENV in
 esac
 ${YAMLES_UTILS} "${args[@]}"
 
+if [ "${MODE}" == "CONFIG" ]
+then
+  # Build archive
+  echo ""
+  echo "============================================================"
+  echo "== Build archive"
+  echo "============================================================"
+  rm "${TMP_PRJ}/META-INF/manifest.mf"
+  tar -C "${TMP_PRJ}" -zcf ${TARGET_ARCHIVE} "."
 
-# Build archive
-echo ""
-echo "============================================================"
-echo "== Build archive"
-echo "============================================================"
-rm "${TMP_PRJ}/META-INF/manifest.mf"
-tar -C "${TMP_PRJ}" -zcf ${TARGET_ARCHIVE} "."
+  # Validate archive
+  echo ""
+  echo "============================================================"
+  echo "== Validate configured project"
+  echo "============================================================"
+  ${YAMLES} validate -s "${TARGET_ARCHIVE}"
 
-# Validate archive
-echo ""
-echo "============================================================"
-echo "== Validate configured project"
-echo "============================================================"
-${YAMLES} validate -s "${TARGET_ARCHIVE}"
-
-echo ""
-echo "============================================================"
-echo "== Finished"
-echo "============================================================"
-echo ""
-echo "Use the following command to deploy the configured archive to"
-echo "the API Gateway."
-echo ""
-echo "  ${AXWAY_BIN}/managedomain --deploy \\"
-echo "    --archive_filename ${TARGET_ARCHIVE} \\"
-echo "    --group APIM \\"
-echo "    --username admin --password changeme"
+  echo ""
+  echo "============================================================"
+  echo "== Finished"
+  echo "============================================================"
+  echo ""
+  echo "Use the following command to deploy the configured archive to"
+  echo "the API Gateway."
+  echo ""
+  echo "  ${AXWAY_BIN}/managedomain --deploy \\"
+  echo "    --archive_filename ${TARGET_ARCHIVE} \\"
+  echo "    --group APIM \\"
+  echo "    --username admin --password changeme"
+fi
